Mostly finished implementation of KeyPair generation.  There is a problem with the SHA3-256 MessageDigest that comes with the JRE which does not have XOF functionality.  Will need to use Keccak library or finish implementing directly.

+++ src/main/java/com/mimiclone/fips203/transforms/NumberTheoretic.java
@@ -11,7 +11,7 @@ public interface NumberTheoretic {
      * @param input An array of 256 integers in modulo q
      * @return An array of 256 integers in modulo q transformed using the NTT algorithm
      */
-    BigInteger[] transform(BigInteger[] input);
+    int[] transform(int[] input);
 
     /**
      * Performs the inverse of a number theoretic transform of an array of 256 integers in modulo q=3329
@@ -20,6 +20,6 @@ public interface NumberTheoretic {
      * @param input An array of 256 integers in modulo q representing a number theoretic transform
      * @return An array of 256 integers in modulo q with the transform reversed
      */
-    BigInteger[] inverse(BigInteger[] input);
+    int[] inverse(int[] input);
 
 }
